Refactor: 커뮤니티 서비스 로직 분리
공통으로 사용되는 메서드들을 새로운 클래스로 분리함

Related to: #31
