+++ src/main/java/pulleydoreurae/careerquestbackend/community/service/PostLikeService.java
@@ -2,7 +2,6 @@ package pulleydoreurae.careerquestbackend.community.service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
@@ -10,14 +9,11 @@ import org.springframework.stereotype.Service;
 
 import lombok.extern.slf4j.Slf4j;
 import pulleydoreurae.careerquestbackend.auth.domain.entity.UserAccount;
-import pulleydoreurae.careerquestbackend.auth.repository.UserAccountRepository;
 import pulleydoreurae.careerquestbackend.community.domain.dto.request.PostLikeRequest;
 import pulleydoreurae.careerquestbackend.community.domain.dto.response.PostResponse;
 import pulleydoreurae.careerquestbackend.community.domain.entity.Post;
 import pulleydoreurae.careerquestbackend.community.domain.entity.PostLike;
-import pulleydoreurae.careerquestbackend.community.repository.CommentRepository;
 import pulleydoreurae.careerquestbackend.community.repository.PostLikeRepository;
-import pulleydoreurae.careerquestbackend.community.repository.PostRepository;
 
 /**
  * 좋아요 Service
@@ -29,19 +25,13 @@ import pulleydoreurae.careerquestbackend.community.repository.PostRepository;
 @Service
 public class PostLikeService {
 
-	private final UserAccountRepository userAccountRepository;
-	private final PostRepository postRepository;
-	private final CommentRepository commentRepository;
 	private final PostLikeRepository postLikeRepository;
+	private final CommonCommunityService commonCommunityService;
 
 	@Autowired
-	public PostLikeService(UserAccountRepository userAccountRepository, PostRepository postRepository,
-			CommentRepository commentRepository,
-			PostLikeRepository postLikeRepository) {
-		this.userAccountRepository = userAccountRepository;
-		this.postRepository = postRepository;
-		this.commentRepository = commentRepository;
+	public PostLikeService(PostLikeRepository postLikeRepository, CommonCommunityService commonCommunityService) {
 		this.postLikeRepository = postLikeRepository;
+		this.commonCommunityService = commonCommunityService;
 	}
 
 	/**
@@ -51,8 +41,8 @@ public class PostLikeService {
 	 * @return 상태변경에 성공하면 true, 실패하면 false
 	 */
 	public boolean changePostLike(PostLikeRequest postLikeRequest) {
-		UserAccount user = findUserAccount(postLikeRequest.getUserId());
-		Post post = findPost(postLikeRequest.getPostId());
+		UserAccount user = commonCommunityService.findUserAccount(postLikeRequest.getUserId());
+		Post post = commonCommunityService.findPost(postLikeRequest.getPostId());
 
 		if (user == null || post == null) {
 			return false;
@@ -64,7 +54,7 @@ public class PostLikeService {
 					.build();
 			postLikeRepository.save(postLike);
 		} else if (postLikeRequest.getIsLiked() == 1) { // 감소
-			PostLike postLike = findPostLike(post, user);
+			PostLike postLike = commonCommunityService.findPostLike(post, user);
 
 			if (postLike == null) {
 				return false;
@@ -82,99 +72,17 @@ public class PostLikeService {
 	 * @return 게시글 리스트
 	 */
 	public List<PostResponse> findAllPostLikeByUserAccount(String userId, Pageable pageable) {
-		UserAccount user = findUserAccount(userId);
+		UserAccount user = commonCommunityService.findUserAccount(userId);
 		List<PostResponse> list = new ArrayList<>();
 
 		postLikeRepository.findAllByUserAccountOrderByIdDesc(user, pageable)
 				.forEach(postLike -> {
 					Post post = postLike.getPost();
-					PostResponse postResponse = PostResponse.builder()
-							.userId(post.getUserAccount().getUserId())
-							.title(post.getTitle())
-							.content(post.getContent())
-							.hit(post.getHit())
-							.commentCount(countComment(post.getId()))
-							.postLikeCount(countPostLike(post.getId()))
-							.category(post.getCategory())
-							.createdAt(post.getCreatedAt())
-							.modifiedAt(post.getModifiedAt())
-							.build();
+					// 게시글 리스트를 반환할땐 좋아요 상태를 사용하지 않는다. (0 으로 지정)
+					PostResponse postResponse = commonCommunityService.postToPostResponse(post, 0);
 					list.add(postResponse);
 				});
 
 		return list;
-	}
-
-	/**
-	 * 회원아이디로 회원정보를 찾아오는 메서드
-	 *
-	 * @param userId 회원아이디
-	 * @return 해당하는 회원정보가 있으면 회원정보를, 없다면 null 리턴
-	 */
-	private UserAccount findUserAccount(String userId) {
-		Optional<UserAccount> findUser = userAccountRepository.findByUserId(userId);
-
-		// 회원정보를 찾을 수 없다면
-		if (findUser.isEmpty()) {
-			log.error("{} 의 회원 정보를 찾을 수 없습니다.", userId);
-			return null;
-		}
-		return findUser.get();
-	}
-
-	/**
-	 * 게시글 id 로 게시글을 찾아오는 메서드
-	 *
-	 * @param postId 게시글 id
-	 * @return 해당하는 게시글이 있다면 게시글을, 없다면 null 리턴
-	 */
-	private Post findPost(Long postId) {
-		Optional<Post> optionalPost = postRepository.findById(postId);
-
-		// 게시글 정보를 찾을 수 없다면
-		if (optionalPost.isEmpty()) {
-			log.error("postId = {} 의 게시글 정보를 찾을 수 없습니다.", postId);
-			return null;
-		}
-		return optionalPost.get();
-	}
-
-	/**
-	 * 게시글과 사용자 정보로 좋아요 정보 가져오기
-	 *
-	 * @param post 게시글 정보
-	 * @param user 사용자 정보
-	 * @return 게시글 좋아요
-	 */
-	private PostLike findPostLike(Post post, UserAccount user) {
-		Optional<PostLike> optionalPostLike = postLikeRepository.findByPostAndUserAccount(post, user);
-
-		if (optionalPostLike.isEmpty()) {
-			log.error("좋아요 정보를 찾을 수 없습니다. post = {},  user = {}", post, user);
-			return null;
-		}
-		return optionalPostLike.get();
-	}
-
-	/**
-	 * 한 게시글에 달린 댓글을 세어주는 메서드
-	 *
-	 * @param postId 게시글 id
-	 * @return 해당 게시글의 댓글 수
-	 */
-	private Long countComment(Long postId) {
-		Post post = findPost(postId);
-		return (long)commentRepository.findAllByPost(post).size();
-	}
-
-	/**
-	 * 한 게시글에 달린 좋아요를 세어주는 메서드
-	 *
-	 * @param postId 게시글 id
-	 * @return 해당 게시글의 좋아요 수
-	 */
-	private Long countPostLike(Long postId) {
-		Post post = findPost(postId);
-		return (long)postLikeRepository.findAllByPost(post).size();
 	}
 }
