+++ src/main/java/pulleydoreurae/careerquestbackend/community/service/PostService.java
@@ -2,12 +2,9 @@ package pulleydoreurae.careerquestbackend.community.service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -15,20 +12,17 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
-import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import pulleydoreurae.careerquestbackend.auth.domain.entity.UserAccount;
-import pulleydoreurae.careerquestbackend.auth.repository.UserAccountRepository;
 import pulleydoreurae.careerquestbackend.common.service.FileManagementService;
 import pulleydoreurae.careerquestbackend.community.domain.dto.request.PostRequest;
 import pulleydoreurae.careerquestbackend.community.domain.dto.response.PostResponse;
 import pulleydoreurae.careerquestbackend.community.domain.entity.Post;
 import pulleydoreurae.careerquestbackend.community.domain.entity.PostImage;
 import pulleydoreurae.careerquestbackend.community.domain.entity.PostViewCheck;
-import pulleydoreurae.careerquestbackend.community.repository.CommentRepository;
 import pulleydoreurae.careerquestbackend.community.repository.PostImageRepository;
 import pulleydoreurae.careerquestbackend.community.repository.PostLikeRepository;
 import pulleydoreurae.careerquestbackend.community.repository.PostRepository;
@@ -45,8 +39,7 @@ import pulleydoreurae.careerquestbackend.community.repository.PostViewCheckRepos
 public class PostService {
 
 	private final PostRepository postRepository;
-	private final UserAccountRepository userAccountRepository;
-	private final CommentRepository commentRepository;
+	private final CommonCommunityService commonCommunityService;
 	private final PostLikeRepository postLikeRepository;
 	private final PostViewCheckRepository postViewCheckRepository;
 	private final PostImageRepository postImageRepository;
@@ -56,13 +49,11 @@ public class PostService {
 	private String IMAGE_PATH;
 
 	@Autowired
-	public PostService(PostRepository postRepository, UserAccountRepository userAccountRepository,
-			CommentRepository commentRepository, PostLikeRepository postLikeRepository,
-			PostViewCheckRepository postViewCheckRepository, PostImageRepository postImageRepository,
-			FileManagementService fileManagementService) {
+	public PostService(PostRepository postRepository, CommonCommunityService commonCommunityService,
+			PostLikeRepository postLikeRepository, PostViewCheckRepository postViewCheckRepository,
+			PostImageRepository postImageRepository, FileManagementService fileManagementService) {
 		this.postRepository = postRepository;
-		this.userAccountRepository = userAccountRepository;
-		this.commentRepository = commentRepository;
+		this.commonCommunityService = commonCommunityService;
 		this.postLikeRepository = postLikeRepository;
 		this.postViewCheckRepository = postViewCheckRepository;
 		this.postImageRepository = postImageRepository;
@@ -76,7 +67,7 @@ public class PostService {
 	 * @return Repository 에서 가져온 리스트 반환
 	 */
 	public List<PostResponse> getPostResponseList(Pageable pageable) {
-		return postListToPostResponseList(postRepository.findAllByOrderByIdDesc(pageable));
+		return commonCommunityService.postListToPostResponseList(postRepository.findAllByOrderByIdDesc(pageable));
 	}
 
 	/**
@@ -88,7 +79,8 @@ public class PostService {
 	 */
 	public List<PostResponse> getPostResponseListByCategory(Long category, Pageable pageable) {
 
-		return postListToPostResponseList(postRepository.findAllByCategoryOrderByIdDesc(category, pageable));
+		return commonCommunityService.postListToPostResponseList(
+				postRepository.findAllByCategoryOrderByIdDesc(category, pageable));
 	}
 
 	/**
@@ -99,11 +91,12 @@ public class PostService {
 	 * @return 회원정보에 맞는 리스트 반환
 	 */
 	public List<PostResponse> getPostListByUserAccount(String userId, Pageable pageable) {
-		UserAccount user = findUserAccount(userId);
+		UserAccount user = commonCommunityService.findUserAccount(userId);
 		if (user == null) {
 			return null;
 		}
-		return postListToPostResponseList(postRepository.findAllByUserAccountOrderByIdDesc(user, pageable));
+		return commonCommunityService.postListToPostResponseList(
+				postRepository.findAllByUserAccountOrderByIdDesc(user, pageable));
 	}
 
 	/**
@@ -118,12 +111,12 @@ public class PostService {
 
 		// 카테고리 없이 전체 검색
 		if (category == null) {
-			return postListToPostResponseList(
+			return commonCommunityService.postListToPostResponseList(
 					postRepository.searchByKeyword(keyword, pageable));
 		}
 
 		// 카테고리 포함 검색
-		return postListToPostResponseList(
+		return commonCommunityService.postListToPostResponseList(
 				postRepository.searchByKeywordAndCategory(keyword, category, pageable));
 	}
 
@@ -134,11 +127,27 @@ public class PostService {
 	 * @return 게시글 dto 를 반환
 	 */
 	public PostResponse findByPostId(HttpServletRequest request, HttpServletResponse response, Long postId) {
-		Post post = findPost(postId);
+		Post post = commonCommunityService.findPost(postId);
 		if (post == null) {
 			return null;
 		}
 
+		String userId = checkView(request, response, postId, post);
+		int isLiked = getIsLiked(userId, post);
+		// 게시글 단건 요청은 게시글 좋아요 정보가 필요하므로 좋아요 정보를 넘기기
+		return commonCommunityService.postToPostResponse(post, isLiked);
+	}
+
+	/**
+	 * 조회수에 대한 처리를 담당하는 메서드
+	 *
+	 * @param request  요청
+	 * @param response 응답
+	 * @param postId   게시글 id
+	 * @param post     게시글 정보
+	 * @return 현재 로그인한 사용자 정보
+	 */
+	public String checkView(HttpServletRequest request, HttpServletResponse response, Long postId, Post post) {
 		// 회원정보(userId) 가져오기
 		String name;
 		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -146,33 +155,29 @@ public class PostService {
 		if (authentication != null && !(authentication instanceof AnonymousAuthenticationToken)) {
 			name = authentication.getName();
 		} else { // 회원정보가 없다면 쿠키와 UUID 를 사용해 식별한다.
-			name = getUUID(request, response);
+			name = commonCommunityService.getUUID(request, response);
 		}
 
-		PostViewCheck postViewCheck = findPostViewCheck(name);
+		PostViewCheck postViewCheck = commonCommunityService.findPostViewCheck(name);
 		if (postViewCheck == null || !postViewCheck.getPostId().equals(postId)) { // Redis 에 저장되어 있지 않다면 저장하고 조회수 증가
 			post.setHit(post.getHit() + 1); // 조회수 증가
 			postViewCheckRepository.save(new PostViewCheck(name, postId)); // 지정한 시간동안 저장
 		}
+		return name;
+	}
 
+	/**
+	 * 좋아요 상태를 반환하는 메서드
+	 *
+	 * @param userId 현재 요청한 사용자 정보
+	 * @param post   게시글 정보
+	 * @return 0이면 좋아요 X, 1이라면 좋아요 상태
+	 */
+	public int getIsLiked(String userId, Post post) {
 		// userId 로 회원을 가져온다.
-		UserAccount user = findUserAccount(name);
-		// null 이거나 좋아요 누른 정보를 가져올 수 없다면 0, 눌렀다면 1
-		int isLiked = postLikeRepository.existsByPostAndUserAccount(post, user) ? 1 : 0;
-
-		return PostResponse.builder()
-				.userId(post.getUserAccount().getUserId())
-				.title(post.getTitle())
-				.content(post.getContent())
-				.images(postImageToStringList(post))
-				.hit(post.getHit())
-				.commentCount(countComment(post.getId()))
-				.postLikeCount(countPostLike(post.getId()))
-				.category(post.getCategory())
-				.isLiked(isLiked)
-				.createdAt(post.getCreatedAt())
-				.modifiedAt(post.getModifiedAt())
-				.build();
+		UserAccount user = commonCommunityService.findUserAccount(userId);
+		// user 가 null 이거나 좋아요 누른 정보를 가져올 수 없다면 0, 눌렀다면 1
+		return postLikeRepository.existsByPostAndUserAccount(post, user) ? 1 : 0;
 	}
 
 	/**
@@ -203,7 +208,7 @@ public class PostService {
 	 */
 	@Transactional // 사진 저장과 게시글 저장을 하나의 트랜잭션으로 묶음
 	public boolean savePost(PostRequest postRequest) {
-		UserAccount user = findUserAccount(postRequest.getUserId());
+		UserAccount user = commonCommunityService.findUserAccount(postRequest.getUserId());
 		List<String> fileNames = postRequest.getImages();
 
 		// 사용자 정보를 확인할 수 없다면 (저장에 실패한다면)
@@ -215,7 +220,7 @@ public class PostService {
 			return false;
 		}
 
-		Post post = postRequestToPost(postRequest, user);
+		Post post = commonCommunityService.postRequestToPost(postRequest, user);
 		try {
 			postRepository.save(post);
 		} catch (Exception e) {
@@ -249,8 +254,8 @@ public class PostService {
 	 * @return 수정에 성공하면 true 실패하면 false
 	 */
 	public boolean updatePost(Long postId, PostRequest postRequest) {
-		Post post = findPost(postId);
-		UserAccount user = findUserAccount(postRequest.getUserId());
+		Post post = commonCommunityService.findPost(postId);
+		UserAccount user = commonCommunityService.findUserAccount(postRequest.getUserId());
 		// 게시글을 찾지못하거나, 회원정보가 없거나(null 인 경우), 작성자와 수정자가 다르다면 실패
 		if (post == null || user == null || !post.getUserAccount().getUserId().equals(user.getUserId())) {
 			return false;
@@ -260,7 +265,7 @@ public class PostService {
 			updateImages(postRequest.getImages(), post);
 		}
 
-		Post updatedPost = postRequestToPostForUpdate(post, postRequest, user);
+		Post updatedPost = commonCommunityService.postRequestToPostForUpdate(post, postRequest, user);
 		postRepository.save(updatedPost);
 
 		return true;
@@ -274,8 +279,8 @@ public class PostService {
 	 * @return 삭제 요청이 성공이면 true 실패하면 false
 	 */
 	public boolean deletePost(Long postId, String userId) {
-		UserAccount user = findUserAccount(userId);
-		Post post = findPost(postId);
+		UserAccount user = commonCommunityService.findUserAccount(userId);
+		Post post = commonCommunityService.findPost(postId);
 		// 게시글을 찾지못하거나, 회원정보가 없거나(null 인 경우), 작성자와 요청자가 다르다면 실패 (권한 없음)
 		if (post == null || user == null || !post.getUserAccount().getUserId().equals(user.getUserId())) {
 			return false;
@@ -284,181 +289,11 @@ public class PostService {
 
 		List<PostImage> fileNames = postImageRepository.findAllByPost(post);
 		// 저장된 사진 파일이 존재한다면 게시글 삭제하면서 사진도 삭제
-		if (fileNames != null) {
+		if (!fileNames.isEmpty()) {
 			fileManagementService.deleteFile(fileNames.stream().map(PostImage::getFileName).toList(), IMAGE_PATH);
 		}
 
 		return true;
-	}
-
-	/**
-	 * 게시글 엔티티 리스트 -> 게시글 response 리스트
-	 *
-	 * @param postList 게시글 엔티티 리스트
-	 * @return 게시글 response 리스트
-	 */
-	private List<PostResponse> postListToPostResponseList(Page<Post> postList) {
-		return postList.stream()
-				.map(post -> PostResponse.builder()
-						.userId(post.getUserAccount().getUserId())
-						.title(post.getTitle())
-						.content(post.getContent())
-						.images(postImageToStringList(post))
-						.hit(post.getHit())
-						.commentCount(countComment(post.getId()))
-						.postLikeCount(countPostLike(post.getId()))
-						.category(post.getCategory())
-						.createdAt(post.getCreatedAt())
-						.modifiedAt(post.getModifiedAt())
-						.build())
-				.toList();
-	}
-
-	/**
-	 * 게시글 id로 게시글을 찾아오는 메서드
-	 *
-	 * @param postId 게시글 id
-	 * @return 게시글이 있다면 게시글을, 없다면 null 리턴
-	 */
-	private Post findPost(Long postId) {
-		Optional<Post> findPost = postRepository.findById(postId);
-
-		if (findPost.isEmpty()) {
-			log.warn("게시글을 찾을 수 없습니다. postId = {}", postId);
-			return null;
-		}
-		return findPost.get();
-	}
-
-	/**
-	 * 회원아이디로 회원정보를 찾아오는 메서드
-	 *
-	 * @param userId 회원아이디
-	 * @return 해당하는 회원정보가 있으면 회원정보를, 없다면 null 리턴
-	 */
-	private UserAccount findUserAccount(String userId) {
-		Optional<UserAccount> findUser = userAccountRepository.findByUserId(userId);
-
-		// 회원정보를 찾을 수 없다면
-		if (findUser.isEmpty()) {
-			log.error("{} 의 회원 정보를 찾을 수 없습니다.", userId);
-			return null;
-		}
-		return findUser.get();
-	}
-
-	/**
-	 * 조회수 중복을 확인하기 위해 Optional 을 제거하는 메서드
-	 *
-	 * @param userId 키값
-	 * @return 제거한 결과 없으면 null, 있다면 해당 객체를 리턴
-	 */
-	private PostViewCheck findPostViewCheck(String userId) {
-		Optional<PostViewCheck> findPostViewCheck = postViewCheckRepository.findById(userId);
-		return findPostViewCheck.orElse(null);
-	}
-
-	/**
-	 * 게시글 요청 -> 게시글 엔티티 변환 메서드 (작성시 사용)
-	 *
-	 * @param postRequest 게시글 요청
-	 * @param user        회원정보
-	 * @return 게시글 엔티티
-	 */
-	private Post postRequestToPost(PostRequest postRequest, UserAccount user) {
-		return Post.builder()
-				.userAccount(user)
-				.title(postRequest.getTitle())
-				.content(postRequest.getContent())
-				.category(postRequest.getCategory())
-				.hit(0L)
-				.build();
-	}
-
-	/**
-	 * 게시글 수정 요청 -> 게시글 엔티티 변환 메서드 (수정시 사용)
-	 *
-	 * @param post        수정할 게시글 엔티티 전달
-	 * @param postRequest 수정할 게시글
-	 * @param user        작성자(수정자)
-	 * @return 게시글 엔티티
-	 */
-	private Post postRequestToPostForUpdate(Post post, PostRequest postRequest, UserAccount user) {
-		// 엔티티의 Setter 사용을 막기위해 새로운 post 생성하며 덮어쓰기
-
-		return Post.builder()
-				.id(post.getId()) // id 를 덮어씌어 수정함
-				.userAccount(user)
-				.title(postRequest.getTitle())
-				.content(postRequest.getContent())
-				.category(postRequest.getCategory())
-				.hit(post.getHit()) // 조회수도 유지
-				.comments(post.getComments()) // 댓글 리스트도 유지
-				.postLikes(post.getPostLikes()) // 좋아요 리스트도 유지
-				.build();
-	}
-
-	/**
-	 * 한 게시글에 달린 댓글을 세어주는 메서드
-	 *
-	 * @param postId 게시글 id
-	 * @return 해당 게시글의 댓글 수
-	 */
-	private Long countComment(Long postId) {
-		Post post = findPost(postId);
-		return (long)commentRepository.findAllByPost(post).size();
-	}
-
-	/**
-	 * 한 게시글에 달린 좋아요를 세어주는 메서드
-	 *
-	 * @param postId 게시글 id
-	 * @return 해당 게시글의 좋아요 수
-	 */
-	private Long countPostLike(Long postId) {
-		Post post = findPost(postId);
-		return (long)postLikeRepository.findAllByPost(post).size();
-	}
-
-	/**
-	 * 회원정보가 없는 사용자를 식별하기 위해 UUID 를 리턴하는 메서드
-	 *
-	 * @param request  요청
-	 * @param response 응답
-	 * @return UUID 값 리턴
-	 */
-	public String getUUID(HttpServletRequest request, HttpServletResponse response) {
-		Cookie[] cookies = request.getCookies();
-		String name = null;
-
-		if (cookies != null) { // 쿠키가 있다면 UUID 가 있는지 확인한다.
-			for (Cookie cookie : cookies) {
-				if (cookie.getName().equals("UUID")) {
-					name = cookie.getValue();
-					break;
-				}
-			}
-		}
-		if (name == null) { // 쿠키가 없거나 있더라도 UUID 값이 없다면 UUID 값을 새로 만들어 쿠키에 저장한다.
-			name = UUID.randomUUID().toString();
-			Cookie cookie = new Cookie("UUID", name);
-			cookie.setMaxAge(60 * 60 * 24); // UUID 쿠키는 10일 동안 사용한다.
-			response.addCookie(cookie); // UUID 값 쿠키에 저장
-		}
-
-		return name;
-	}
-
-	/**
-	 * 한 게시글달린 사진 파일명들을 반환하는 메서드
-	 *
-	 * @param post 게시글 정보
-	 * @return 저장된 파일명 리스트
-	 */
-	private List<String> postImageToStringList(Post post) {
-		return postImageRepository.findAllByPost(post).stream()
-				.map(PostImage::getFileName)
-				.toList();
 	}
 
 	/**
